pyproject.toml
[tool.poetry]
name = "gcp-image-vision"
version = "0.1.0"
description = "Google Cloud Platform Image Vision API integration"
authors = ["Yann LEGER <yann@koyeb.com>"]
license = "Apache License 2.0"

[tool.poetry.dependencies]
python = "^3.8"
google-cloud-vision = "^1.0.0"
boto3 = "^1.14.57"

[tool.poetry.dev-dependencies]
tox = "^3.20.0"
black = "^20.8b1"
isort = "^5.5.1"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
